% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Sigmoid Function}
\usage{
expit(x)
}
\arguments{
\item{x}{a variable}
}
\description{
This function applies sigmoid function on X
}
\keyword{expit}
\keyword{sigmoid}
