% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{LinearRegression}
\alias{LinearRegression}
\title{Linear Regression Function}
\usage{
LinearRegression(X, Y)
}
\arguments{
\item{X}{n x p matrix of explanatory variables}

\item{Y}{n dimensional vector of responses}
}
\description{
This function performs linear regression of Y on X.
}
\keyword{Linear}
\keyword{Regression}
