% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM_error.R
\name{LM_error}
\alias{LM_error}
\title{Linear Regression Error}
\usage{
LM_error(X, Y, my_coef)
}
\arguments{
\item{X}{an n x p matrix of explanatory variables}

\item{Y}{Output variables}

\item{my_coef}{reponse of linear regression}
}
\description{
This function calculates Error of linear regression
}
\keyword{LR}
\keyword{Linear}
\keyword{Regression}
\keyword{error}
