% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_error.R
\name{log_error}
\alias{log_error}
\title{Logistic Error}
\usage{
log_error(X, my_coef)
}
\arguments{
\item{X}{an n x p matrix of explanatory variables}

\item{my_coef}{reponse of logistic regression}
}
\description{
This function calculates Error of logistic regression
}
\keyword{LR}
\keyword{Logistic}
\keyword{Regression}
\keyword{error}
