% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticRegression.R
\name{LogisticRegression}
\alias{LogisticRegression}
\title{Logistic Regression Function}
\usage{
LogisticRegression(X, Y)
}
\arguments{
\item{X}{n x p matrix of explanatory variables}

\item{Y}{n dimensional vector of binary responses}
}
\description{
This function performs logistic regression of Y on X.
}
\keyword{Logistic}
\keyword{Regression}
