% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{PCA Function}
\usage{
PCA(A, numIter = 1000)
}
\arguments{
\item{A}{square matrix}

\item{numIter}{number of iterations}
}
\description{
This function performs PCA on matrix A using QR function, QR.
}
\keyword{PCA}
\keyword{eigen}
