% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR.R
\name{QR}
\alias{QR}
\title{QR Function}
\usage{
QR(A)
}
\arguments{
\item{A}{an n x m matrix}
}
\description{
This function performs QR decomposition on the matrix A
}
\keyword{QR}
\keyword{decomposition}
